--- iceberg/common/src/main/java/com/nvidia/spark/rapids/fileio/iceberg/BaseIcebergInputFile.java
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2026, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nvidia.spark.rapids.fileio.iceberg;
-
-import com.nvidia.spark.rapids.jni.fileio.RapidsInputFile;
-import com.nvidia.spark.rapids.jni.fileio.SeekableInputStream;
-import org.apache.iceberg.io.InputFile;
-
-import java.io.IOException;
-import java.util.Objects;
-
-/** Base class for {@link RapidsInputFile} implementations backed by an Iceberg input file. */
-public abstract class BaseIcebergInputFile implements RapidsInputFile {
-  private final InputFile delegate;
-
-  protected BaseIcebergInputFile(InputFile delegate) {
-    Objects.requireNonNull(delegate, "delegate can't be null");
-    this.delegate = delegate;
-  }
-
-  @Override
-  public String path() {
-    return delegate.location();
-  }
-
-  @Override
-  public long getLength() throws IOException {
-    return delegate.getLength();
-  }
-
-  @Override
-  public SeekableInputStream open() throws IOException {
-    return new IcebergInputStream(delegate.newStream());
-  }
-
-  /**
-   * Returns the underlying Iceberg InputFile delegate.
-   *
-   * @return the Iceberg InputFile delegate
-   */
-  public InputFile getDelegate() {
-    return delegate;
-  }
-}
