+++ iceberg/common/src/main/java/com/nvidia/spark/rapids/fileio/iceberg/IcebergFileIO.java
@@ -49,7 +49,7 @@ public class IcebergFileIO implements RapidsFileIO {
 
 
   @Override
-  public BaseIcebergInputFile newInputFile(String path) throws IOException {
+  public IcebergInputFile newInputFile(String path) throws IOException {
     return newInputFile(delegate.newInputFile(path));
   }
 
@@ -58,7 +58,7 @@ public class IcebergFileIO implements RapidsFileIO {
    *
    * @param inputFile the Iceberg input file to wrap
    */
-  public BaseIcebergInputFile newInputFile(InputFile inputFile) {
+  public IcebergInputFile newInputFile(InputFile inputFile) {
     Objects.requireNonNull(inputFile, "inputFile can't be null");
     return IcebergS3InputFile.maybeCreate(inputFile, delegate);
   }
