+++ iceberg/iceberg-1-10-x/src/main/java/com/nvidia/spark/rapids/iceberg/iceberg110x/ShimUtilsImpl.java
@@ -18,7 +18,7 @@ package com.nvidia.spark.rapids.iceberg.iceberg110x;
 
 import com.nvidia.spark.rapids.GpuMetric;
 import com.nvidia.spark.rapids.RapidsConf;
-import com.nvidia.spark.rapids.fileio.iceberg.BaseIcebergInputFile;
+import com.nvidia.spark.rapids.fileio.iceberg.IcebergInputFile;
 import com.nvidia.spark.rapids.iceberg.IcebergShimUtils;
 import org.apache.hadoop.fs.Path;
 import org.apache.iceberg.*;
@@ -72,7 +72,7 @@ public class ShimUtilsImpl implements IcebergShimUtils {
 
     @Override
     public ParquetFileReader openParquetReader(
-            BaseIcebergInputFile inputFile,
+            IcebergInputFile inputFile,
             Path filePath,
             ParquetReadOptions options,
             scala.collection.immutable.Map<String, GpuMetric> metrics) throws IOException {
